+++ mmdet/models/dense_heads/anchor_head.py
@@ -11,7 +11,7 @@ from mmdet.core import (anchor_inside_flags, build_assigner, build_bbox_coder,
 from ..builder import HEADS, build_loss
 from .base_dense_head import BaseDenseHead
 from .dense_test_mixins import BBoxTestMixin
-
+from utils import int_dtype
 
 @HEADS.register_module()
 class AnchorHead(BaseDenseHead, BBoxTestMixin):
@@ -255,7 +255,7 @@ class AnchorHead(BaseDenseHead, BBoxTestMixin):
         bbox_weights = torch.zeros_like(anchors)
         labels = anchors.new_full((num_valid_anchors, ),
                                   self.num_classes,
-                                  dtype=torch.long)
+                                  dtype=int_dtype)
         label_weights = anchors.new_zeros(num_valid_anchors, dtype=torch.float)
 
         pos_inds = sampling_result.pos_inds
